+++ CHANGELOG.md
@@ -1,3 +1,4 @@
-## v1.1.5 (1.20.1)
+## v1.1.6 (1.20.1)
 - added new expiremental Alt-render
 - Altrenderer should be fully compatible with Iris, Oculus and Optifine
+- This Altrenderer can be toggled in the miapi.json config  the default Option is AUTO, use DEFAULT_RENDERER or FALLBACK_RENDERER in case any issues are encounted. Please report any encountered issues with the Alt Renderer